Refactor 12, start on 13.
The solution to challenge 12 is faster now, and hopefully more readable,
but I left some debug output in because it's still pretty slow.

Tried to make the oracle code make more sense. It's no longer necessary
to pass in the same key manually each time. Rust closures are weird.

Also started on challenge 13.