Fix false-positives in challenge 17
When the server claims that a message has valid padding, we don't know
the actual padding byte for sure. Double-check that the padding is still
valid regardless of the value of the preceding byte to establish an
upper-bound on the padding byte value. This avoids making the wrong
decision in potentially ambiguous cases where e.g. byte 15 is 0x02,
resulting in a "valid padding" message when byte 16 is 0x01 or 0x02